Move sanity checks to separate file
